Add user lan link state monitoring stask
This task will be started by MotoROS2 as part of its connect-to-micro-ROS-Agent sequence, but only if configured to do so.

Whenever this task detects the ethernet port it's monitoring goes down, it forces MotoROS2's shutdown behaviour by setting the global AgentIsConnected to false. Doing this will also shutdown the monitoring task (as there is now no longer a reason to monitor the port), which will be restarted again if/when MotoROS2 detects it can reconnect to the Agent (because the network port came back up).