Avoid printing both sent and receive stamp in the logger
PR #61 added a timestamp to the log message sent out by MotoROS2. Use that timestamp instead of calculating a new one upon reception.

The format of the timestamp prefixed by MotoROS2 is exactly the same as what `debug_listener.py` used to prefix. But due to limitations on the VxWorks side, will likely not take time-zones into account.