Raise alarm in case none of the calib files loaded
Note: as mentioned in the comments, it's perfectly OK for multi-group
systems to not have been calibrated. Whether or not calibration
succeeded to load is therefore not taken into account when determining
whether controller initialisation was successful or not (`bInitOk`).

But, in case TF broadcasts are enabled, uncalibrated groups
(specifically: robots) will have their origins coincident, which will
(most likely) not be representative of their physical configuration (ie:
in case those groups are different manipulators, those will not occupy
the same physical space).

In those cases an alarm is raised to notify the user of this potential
misconfiguration.

But only if the alarm hasn't been disabled by user configuration.