ci: test only specific upstream M+ uROS releases
instead of the full Cartesian product of all matrix lists.

Upstream (ie: micro_ros_motoplus) may not support all combinations of
all versions of micro-ROS on all controller platforms for specific
releases, hence the need to be able to exhaustively list which specific
releases should be included in the build.