Prevent building FJT tests stand-alone
they call static functions in `ActionServer_FJT.c` and thus can't be
built stand-alone.

As `Tests_ActionServer_FJT.c` is part of the VS project, mpBuilder will
attempt to build it stand-alone, which leads to a failed build.

Use the pre-processor to present an empty file to mpBuilder if it tries
to do that.